+++ src/check.rs
@@ -445,7 +445,7 @@ impl<'a> VirtualEnv<'a> {
                     let f = self.pop_func()?;
                     self.handle_sig(f.signature())?;
                 }
-                Unpack | Memo | Comptime => {
+                Unpack | Content | Memo | Comptime => {
                     let f = self.pop_func()?;
                     self.handle_sig(f.signature())?;
                 }