+++ src/lex.rs
@@ -608,10 +608,6 @@ impl<'a> Lexer<'a> {
                 "≅" => self.end(Primitive::Match, start),
                 "∶" => self.end(Primitive::Flip, start),
                 "⍘" => self.end(Primitive::Un, start),
-                "⊔" => {
-                    self.end(Primitive::Un, self.loc);
-                    self.end(Primitive::Box, start);
-                }
                 "⊝" => self.end(Primitive::Deduplicate, start),
 
                 "(" => self.end(OpenParen, start),