+++ src/primitive/defs.rs
@@ -705,13 +705,13 @@ primitive!(
     /// [under][un][box] works because `un``un``box` is just `box`. For each element, it [un][box]es the array out, does something to it, then [box]es the result.
     /// ex: .{1_2_3 4_5 [7]}
     ///   : ∵⍜°□(⬚0↙3)
-    /// If you do not need to re-[box] the result, you can use [unpack] instead.
-    /// [unpack] implicitly [un][box]es all box elements that are passed to a function, but does not re-[box] results.
+    /// If you do not need to re-[box] the result, you can use [content] instead.
+    /// [content] [un][box]es all box elements that are passed to a function before calling it.
     /// ex: {1_2_3 9_2 5_5_5_5}
-    ///   : ⊐≡/+
+    ///   : ≡⊔/+
     /// This is the main way to [join] a list of [box]ed strings.
-    /// ex: ⊐/⊂       {"Join" "these" "strings"}
-    /// ex: ⊐/(⊂⊂:@ ) {"Join" "these" "strings"}
+    /// ex: /⊔⊂       {"Join" "these" "strings"}
+    /// ex: /⊔(⊂⊂:@ ) {"Join" "these" "strings"}
     (1, Box, MonadicArray, ("box", '□')),
     /// Parse a string as a number
     ///
@@ -1229,6 +1229,13 @@ primitive!(
     /// ex: ⊐(¯□3) # Used
     /// ex: ⊐( □3) # Not used
     ([1], Unpack, OtherModifier, ("unpack", '⊐')),
+    /// Unbox the arguments to a function before calling it
+    ///
+    /// ex: ⊔⊂ □[1 2 3] □[4 5 6]
+    ///   :  ⊂ □[1 2 3] □[4 5 6]
+    /// A common use of [content] is to collapse a list of [box]ed arrays with [reduce].
+    /// ex: /⊔⊂ {1_2_3 4_5 6}
+    ([1], Content, OtherModifier, ("content", '⊔')),
     /// Discard the top stack value then call a function
     ///
     /// See the [Advanced Stack Manipulation Tutorial](/docs/advancedstack) for a more complete understanding of why [gap] is useful.