+++ src/run.rs
@@ -988,9 +988,13 @@ code:
         self.rt.function_stack.clear();
         take(&mut self.rt.stack)
     }
-    /// Return a reference to the stack
-    pub fn get_stack(&self) -> &[Value] {
+    /// Get a reference to the stack
+    pub fn stack(&self) -> &[Value] {
         &self.rt.stack
+    }
+    /// Get a mutable reference to the stack data
+    pub fn stack_mut(&mut self) -> &mut [Value] {
+        &mut self.rt.stack
     }
     /// Pop a function from the function stack
     pub fn pop_function(&mut self) -> UiuaResult<Function> {